line_tools/requirements.txt
Flask>=2.0
line-bot-sdk>=2.0
gunicorn>=20.0
